@@ -16,7 +16,11 @@ import (
 )
 
 func main() {
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelInfo}))
+	logOpts := &slog.HandlerOptions{Level: slog.LevelInfo}
+	logger := slog.New(slog.NewMultiHandler(
+		slog.NewTextHandler(os.Stdout, logOpts),
+		slog.NewJSONHandler(os.Stderr, logOpts),
+	))
 	slog.SetDefault(logger)
 
 	port := envOrDefault("PORT", "8080")
